+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file. The format 
 
 ## Unreleased
 ### Fixed
-- Default Panel background ignored when preset present
+- Default Panel background ignored when a preset was present
+- Panel now checks presets for a custom background before falling back
 
 ## [0.16.0]
 - Adjust sizing and spacing for: